chore(structure): temp rename folder to trigger case-sensitive rename
