+++ src/adapte/book-adapter-interface.ts
