+++ src/middlewar/global-error-handler.ts
@@ -1,5 +1,5 @@
 import { ExpressErrorMiddlewareInterface, Middleware } from "routing-controllers";
-import { CustomError } from "../custom-errors/custom-erros";
+import { CustomError } from "../custom-error/custom-erros";
 import { Request, Response, NextFunction } from "express";
 import { Service } from "typedi";
